requirements.txt
flask
torch
torchvision
timm
numpy
pillow